Lab4Update

+++ Controllers/BooksController.cs
@@ -19,11 +19,11 @@ namespace Timu_Vlad_Lab2.Controllers
             _context = context;
         }
 
-        // GET: Books
-        public async Task<IActionResult> Index()
-        {
-            return View(await _context.Books.ToListAsync());
-        }
+        /* // GET: Books
+         *//*public async Task<IActionResult> Index()
+         {
+             return View(await _context.Books.ToListAsync());
+         }*/
 
         // GET: Books/Details/5
         public async Task<IActionResult> Details(int? id)
@@ -34,6 +34,9 @@ namespace Timu_Vlad_Lab2.Controllers
             }
 
             var book = await _context.Books
+                .Include(s => s.Orders)
+                .ThenInclude(e => e.Customer)
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
             if (book == null)
             {
@@ -54,13 +57,20 @@ namespace Timu_Vlad_Lab2.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,Title,Author,Price")] Book book)
+        public async Task<IActionResult> Create([Bind("Title,Author,Price")] Book book)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(book);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(book);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+            }
+            catch (DbUpdateException ex)
+            {
+                ModelState.AddModelError("", "Unable to save changes. " + "Try again, and if the problem persists ");
             }
             return View(book);
         }
@@ -84,40 +94,34 @@ namespace Timu_Vlad_Lab2.Controllers
         // POST: Books/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,Title,Author,Price")] Book book)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != book.ID)
+            if (id == null)
             {
                 return NotFound();
             }
+            var studentToUpdate = await _context.Books.FirstOrDefaultAsync(s => s.ID == id);
 
-            if (ModelState.IsValid)
+            if (await TryUpdateModelAsync<Book>(studentToUpdate, "", s => s.Author, s => s.Title, s => s.Price))
             {
                 try
                 {
-                    _context.Update(book);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException ex)
                 {
-                    if (!BookExists(book.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    ModelState.AddModelError("", "Unable to save changes. " + "Try again, and if the problem persists");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            return View(book);
+            return View(studentToUpdate);
         }
 
+
         // GET: Books/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -125,10 +129,15 @@ namespace Timu_Vlad_Lab2.Controllers
             }
 
             var book = await _context.Books
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
             if (book == null)
             {
                 return NotFound();
+            }
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] = "Delete failed. Try again";
             }
 
             return View(book);
@@ -140,14 +149,65 @@ namespace Timu_Vlad_Lab2.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var book = await _context.Books.FindAsync(id);
-            _context.Books.Remove(book);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
+            if (book == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            try
+            {
+                _context.Books.Remove(book);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException ex)
+            {
 
-        private bool BookExists(int id)
-        {
-            return _context.Books.Any(e => e.ID == id);
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
+        public async Task<IActionResult> Index(
+            string sortOrder,
+            string currentFilter,
+            string searchString,
+            int? pageNumber)
+    {
+        ViewData["CurrentSort"] = sortOrder;
+        ViewData["TitleSortParm"] = String.IsNullOrEmpty(sortOrder) ? "title_desc" : "";
+        ViewData["PriceSortParm"] = sortOrder == "Price" ? "price_desc" : "Price";
+        if (searchString != null)
+        {
+            pageNumber = 1;
+        }
+        else
+        {
+            searchString = currentFilter;
+        }
+        ViewData["CurrentFilter"] = searchString;
+        var books = from b in _context.Books
+                    select b;
+        if (!String.IsNullOrEmpty(searchString))
+        {
+            books = books.Where(s => s.Title.Contains(searchString));
+        }
+        switch (sortOrder)
+        {
+            case "title_desc":
+                books = books.OrderByDescending(b => b.Title);
+                break;
+            case "Price":
+                books = books.OrderBy(b => b.Price);
+                break;
+            case "price_desc":
+                books = books.OrderByDescending(b => b.Price);
+                break;
+            default:
+                books = books.OrderBy(b => b.Title);
+                break;
+        }
+        int pageSize = 2;
+        return View(await PaginatedList<Book>.CreateAsync(books.AsNoTracking(), pageNumber ??
+       1, pageSize));
     }
+    }
+  
 }
